# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MAROC EXPLORER API",
        "description": "This is the API documentation for the MAROC EXPLORER API. It handles webhook requests and responses, allowing users to generate and display data dynamically.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/destinations": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Display a listing of all destinations",
                "description": "Returns a list of all destinations with their associated activities and dishes.",
                "operationId": "27fd12df2befee89549ebd10d33bae69",
                "responses": {
                    "200": {
                        "description": "Destinations fetched successfully."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Store a new destination",
                "description": "Creates a new destination and returns it.",
                "operationId": "9e07e09ca0eb956a07abd3ec5261db1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "itineraire_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chefchaouen"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The Blue City"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "chefchaouen.jpg",
                                        "nullable": true
                                    },
                                    "itineraire_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Destination created successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/destinations/{id}": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Display a specific destination",
                "description": "Returns a specific destination with its associated activities and dishes.",
                "operationId": "d80273c09a14d27732106d2dd4fd870a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the destination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Destination fetched successfully."
                    },
                    "404": {
                        "description": "Destination not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Update an existing destination",
                "description": "Updates a specific destination and returns it.",
                "operationId": "aa3106e5ea76ffc863ffde4db28bc0e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the destination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chefchaouen Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The Blue City Updated"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "chefchaouen_new.jpg",
                                        "nullable": true
                                    },
                                    "itineraire_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Destination updated successfully."
                    },
                    "404": {
                        "description": "Destination not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/destinations/search/{query}": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Search destinations",
                "description": "Search for destinations by name.",
                "operationId": "c90e3855b473843fe2194fec1db6814f",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "The search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results fetched successfully."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}